# Optimize ORC reader performance for list data