--- a/cpp/src/io/orc/reader_impl.cu
+++ b/cpp/src/io/orc/reader_impl.cu
@@ -722,12 +722,13 @@ struct list_buffer_data {
 };
 
 // Generates offsets for list buffer from number of elements in a row.
-void generate_offsets_for_list(host_span<list_buffer_data> const& buff_data,
-                               rmm::cuda_stream_view stream)
+void generate_offsets_for_list(host_span<list_buffer_data> buff_data, rmm::cuda_stream_view stream)
 {
   for (auto& list_data : buff_data) {
-    thrust::exclusive_scan(
-      rmm::exec_policy(stream), list_data.data, list_data.data + list_data.size, list_data.data);
+    thrust::exclusive_scan(rmm::exec_policy_nosync(stream),
+                           list_data.data,
+                           list_data.data + list_data.size,
+                           list_data.data);
   }
 }
 
@@ -1332,7 +1333,7 @@ table_with_metadata reader::impl::read(uint64_t skip_rows,
           }
         });
 
-      if (buff_data.size()) { generate_offsets_for_list(buff_data, _stream); }
+      if (not buff_data.empty()) { generate_offsets_for_list(buff_data, _stream); }
     }
   }