## Description
<!-- Provide a standalone description of changes in this PR. -->
<!-- Reference any issues closed by this PR with "closes #1234". -->
<!-- Note: The pull request title will be included in the CHANGELOG. -->
For list types the ORC reader needs to generate offsets from the sizes of nested lists. This process was previously being parallelized over columns. In practice even with wide tables we have enough rows that parallelizing over rows always makes more sense, so this PR swaps the parallelization strategy. Here are some relevant benchmark results

<details><summary>Test script from #13674</summary>
<p>

```
# Old implementation
parquet (numrows 1000) 0.0062103271484375 seconds                                                                                                                                                          
orc (numrows 1000) 0.0033025741577148438 seconds                                                                                                                                                           
parquet (numrows 50000000) 0.016933441162109375 seconds                                                                                                                                                    
orc (numrows 50000000) 1.1314363479614258 seconds                      

# New implementation
parquet (numrows 1000) 0.0062923431396484375 seconds                                                                                                                                                       
orc (numrows 1000) 0.003666400909423828 seconds                                                                                                                                                            
parquet (numrows 50000000) 0.016930341720581055 seconds                                                                                                                                                    
orc (numrows 50000000) 0.045706748962402344 seconds    
```
</p>
</details>
The changes clearly close the gap for the above case.

<details><summary>libcudf ORC reader benchmarks</summary>
<p>

```
# Old implementation
  |            LIST | DEVICE_BUFFER |           0 |          1 |     10x |  52.919 ms | 0.06% |  52.913 ms | 0.06% |      10146288118 |         1.005 GiB |       488.481 MiB |                            
  |            LIST | DEVICE_BUFFER |        1000 |          1 |      6x |  87.514 ms | 0.45% |  87.508 ms | 0.45% |       6135114527 |         1.014 GiB |       363.095 MiB |                            
  |            LIST | DEVICE_BUFFER |           0 |         32 |      8x |  67.902 ms | 0.12% |  67.897 ms | 0.12% |       7907179514 |       574.083 MiB |        34.179 MiB |                            
  |            LIST | DEVICE_BUFFER |        1000 |         32 |      8x |  70.422 ms | 0.09% |  70.416 ms | 0.09% |       7624239955 |       572.610 MiB |        23.229 MiB |    


# New implementation
  |            LIST | DEVICE_BUFFER |           0 |          1 |     12x |  45.033 ms | 0.09% |  45.028 ms | 0.09% |      11923067760 |         1.005 GiB |       488.481 MiB |                            
  |            LIST | DEVICE_BUFFER |        1000 |          1 |      7x |  79.599 ms | 0.49% |  79.593 ms | 0.49% |       6745162704 |         1.014 GiB |       363.095 MiB |                            
  |            LIST | DEVICE_BUFFER |           0 |         32 |      9x |  60.006 ms | 0.12% |  60.000 ms | 0.12% |       8947826725 |       574.083 MiB |        34.179 MiB |                            
  |            LIST | DEVICE_BUFFER |        1000 |         32 |      8x |  62.578 ms | 0.18% |  62.573 ms | 0.18% |       8579974589 |       572.610 MiB |        23.229 MiB |   

```
</p>
</details>
The ORC reader benchmarks above show a marked improvement for the benchmarked list data. All other benchmarks in that file remain unchanged.

To also test the behavior of the new code on wide tables, I reran the test script from #13674, except instead of using tall tables I used wide tables containing varying number of columns.

<details><summary>Wide table</summary>
<p>

```
# Old code
orc (numrows 1, numcols 10) 0.004950284957885742 seconds                                                                                                                                                              
orc (numrows 1, numcols 100) 0.012974739074707031 seconds                                                       
orc (numrows 1, numcols 1000) 0.10814499855041504 seconds                                                                                                                                                             
orc (numrows 1, numcols 4000) 0.4879453182220459 seconds                                                        
orc (numrows 1, numcols 5963) 0.7366838455200195 seconds    

orc (numrows 1000, numcols 10) 0.005819559097290039 seconds
orc (numrows 1000, numcols 100) 0.013351678848266602 seconds
orc (numrows 1000, numcols 1000) 0.10951924324035645 seconds
orc (numrows 1000, numcols 4000) 0.4978141784667969 seconds
orc (numrows 1000, numcols 5963) 0.7526876926422119 seconds

orc (numrows 10000, numcols 10) 0.004932880401611328 seconds
orc (numrows 10000, numcols 100) 0.014151573181152344 seconds
orc (numrows 10000, numcols 1000) 0.1169898509979248 seconds
orc (numrows 10000, numcols 4000) 0.5291402339935303 seconds
orc (numrows 10000, numcols 5963) 0.7931733131408691 seconds

orc (numrows 100000, numcols 10) 0.007727384567260742 seconds
orc (numrows 100000, numcols 100) 0.023123741149902344 seconds
orc (numrows 100000, numcols 1000) 0.19863557815551758 seconds
orc (numrows 100000, numcols 4000) 0.8940346240997314 seconds
orc (numrows 100000, numcols 5963) 1.4564213752746582 seconds

# New code
orc (numrows 1, numcols 10) 0.0052776336669921875 seconds
orc (numrows 1, numcols 100) 0.014220952987670898 seconds
orc (numrows 1, numcols 1000) 0.12338590621948242 seconds
orc (numrows 1, numcols 4000) 0.5486130714416504 seconds
orc (numrows 1, numcols 5963) 0.8224778175354004 seconds

orc (numrows 1000, numcols 10) 0.04096388816833496 seconds
orc (numrows 1000, numcols 100) 0.014076948165893555 seconds
orc (numrows 1000, numcols 1000) 0.16178202629089355 seconds
orc (numrows 1000, numcols 4000) 0.5019838809967041 seconds
orc (numrows 1000, numcols 5963) 0.7861366271972656 seconds

orc (numrows 10000, numcols 10) 0.043140411376953125 seconds
orc (numrows 10000, numcols 100) 0.015340805053710938 seconds
orc (numrows 10000, numcols 1000) 0.1310873031616211 seconds
orc (numrows 10000, numcols 4000) 0.576120138168335 seconds
orc (numrows 10000, numcols 5963) 0.8083529472351074 seconds

orc (numrows 100000, numcols 10) 0.059877634048461914 seconds
orc (numrows 100000, numcols 100) 0.02300548553466797 seconds
orc (numrows 100000, numcols 1000) 0.21386289596557617 seconds
orc (numrows 100000, numcols 4000) 0.9530372619628906 seconds
orc (numrows 100000, numcols 5963) 1.4837677478790283 seconds
```
</p>
</details>

Depending on how important it is for us to preserve the old performance for wide tables we could dispatch to either the old implementation or the new one conditionally depending on some heuristic.

Note that I also removed what appears to be an unnecessary stream synchronization. That likely won't affect performance in any microbenchmarks but is worthwhile in case it helps improve asynchronous execution overall.

Resolves #13674 

## Checklist
- [x] I am familiar with the [Contributing Guidelines](https://github.com/rapidsai/cudf/blob/HEAD/CONTRIBUTING.md).
- [x] New or existing tests cover these changes.
- [x] The documentation is up to date with these changes.